Fix proxy example in sdk-reference/go.md
The HTTP Proxy example did not work "out-of-the-box" as it was written
previously. The `url.Parse()` function returns two values, which means
it cause a "too many arguments" error when used inline like this, and
the missing commas returns "syntax" errors when omitted.

This change makes it possible to just copy-paste this example and have
it function directly.